# docs/swagger.yaml
basePath: /api/v1
...
host: leti-backend.onrender.com
info:
  contact: {}
  description: Leti Server REST API
  title: Leti API
  version: "1.0"
paths:
...
  /bookings/{id}/pay:
    post:
      consumes:
      - application/json
      description: Initiates payment for a booking that has been confirmed by the
        artisan. Supports wallet (instant) and Paystack (redirect). The full amount
        is held in escrow and only released to the artisan after the booking is marked
        completed. Only callable by the client on the booking.
      parameters:
      - description: Booking UUID
        in: path
        name: id
        required: true
        type: string
      - description: 'payment_method: wallet | paystack'
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            payment_method:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              commission:
                type: number
              escrow_id:
                type: string
              message:
                type: string
              payout:
                type: number
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "402":
          description: Payment Required
          schema:
            properties:
              available:
                type: number
              code:
                type: string
              error:
                type: string
              required:
                type: number
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pay for a confirmed booking
      tags:
      - Booking
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"